Move more helpers from DomainAssembly to Assembly
