+++ src/coreclr/debug/daccess/dacdbiimpl.cpp
@@ -672,7 +672,7 @@ void DacDbiInterfaceImpl::GetCompilerFlags (
     }
 
     // Get the underlying module - none of this is AppDomain specific
-    Module * pModule = pDomainAssembly->GetModule();
+    Module * pModule = pDomainAssembly->GetAssembly()->GetModule();
     DWORD dwBits = pModule->GetDebuggerInfoBits();
     *pfAllowJITOpts = !CORDisableJITOptimizations(dwBits);
     *pfEnableEnC = pModule->IsEditAndContinueEnabled();
@@ -721,7 +721,7 @@ HRESULT DacDbiInterfaceImpl::SetCompilerFlags(VMPTR_DomainAssembly vmDomainAssem
 
     DWORD        dwBits      = 0;
     DomainAssembly * pDomainAssembly = vmDomainAssembly.GetDacPtr();
-    Module *     pModule     = pDomainAssembly->GetModule();
+    Module *     pModule     = pDomainAssembly->GetAssembly()->GetModule();
     HRESULT      hr          = S_OK;
 
 
@@ -821,7 +821,7 @@ SIZE_T DacDbiInterfaceImpl::GetArgCount(MethodDesc * pMD)
         return 0;
     }
 
-    MetaSig msig(pCallSig, cbCallSigSize, pMD->GetModule(), NULL, MetaSig::sigMember);
+    MetaSig msig(pCallSig, cbCallSigSize, pMD->GetAssembly()->GetModule(), NULL, MetaSig::sigMember);
 
     // Get the arg count.
     UINT32 NumArguments = msig.NumFixedArgs();
@@ -1019,7 +1019,7 @@ void DacDbiInterfaceImpl::GetSequencePoints(MethodDesc *     pMethodDesc,
 #endif
         // if there is a profiler load-time mapping and not a rejit mapping, apply that instead
         InstrumentedILOffsetMapping loadTimeMapping =
-            pMethodDesc->GetModule()->GetInstrumentedILOffsetMapping(pMethodDesc->GetMemberDef());
+            pMethodDesc->GetAssembly()->GetModule()->GetInstrumentedILOffsetMapping(pMethodDesc->GetMemberDef());
         ComposeMapping(&loadTimeMapping, mapCopy, &entryCount);
 #ifdef FEATURE_REJIT
     }
@@ -1091,7 +1091,7 @@ void DacDbiInterfaceImpl::GetILCodeAndSig(VMPTR_DomainAssembly vmDomainAssembly,
     DD_ENTER_MAY_THROW;
 
     DomainAssembly * pDomainAssembly = vmDomainAssembly.GetDacPtr();
-    Module *     pModule     = pDomainAssembly->GetModule();
+    Module *     pModule     = pDomainAssembly->GetAssembly()->GetModule();
     RVA          methodRVA   = 0;
     DWORD        implFlags;
 
@@ -1330,7 +1330,7 @@ void DacDbiInterfaceImpl::GetNativeCodeInfo(VMPTR_DomainAssembly         vmDomai
     pCodeInfo->Clear();
 
     DomainAssembly * pDomainAssembly = vmDomainAssembly.GetDacPtr();
-    Module *     pModule     = pDomainAssembly->GetModule();
+    Module *     pModule     = pDomainAssembly->GetAssembly()->GetModule();
 
     MethodDesc* pMethodDesc = FindLoadedMethodRefOrDef(pModule, functionToken);
     pCodeInfo->vmNativeCodeMethodDescToken.SetHostPtr(pMethodDesc);
@@ -2866,7 +2866,7 @@ TypeHandle DacDbiInterfaceImpl::GetClassOrValueTypeHandle(DebuggerIPCE_BasicType
     else
     {
         DomainAssembly * pDomainAssembly = pData->vmDomainAssembly.GetDacPtr();
-        Module *     pModule = pDomainAssembly->GetModule();
+        Module *     pModule = pDomainAssembly->GetAssembly()->GetModule();
 
         typeHandle = ClassLoader::LookupTypeDefOrRefInModule(pModule, pData->metadataToken);
         if (typeHandle.IsNull())
@@ -3842,7 +3842,7 @@ FieldDesc * DacDbiInterfaceImpl::GetEnCFieldDesc(const EnCHangingFieldInfo * pEn
         FieldDesc * pFD = NULL;
 
         DomainAssembly * pDomainAssembly = pEnCFieldInfo->GetObjectTypeData().vmDomainAssembly.GetDacPtr();
-        Module     * pModule     = pDomainAssembly->GetModule();
+        Module     * pModule     = pDomainAssembly->GetAssembly()->GetModule();
 
         // get the type handle for the object
         TypeHandle typeHandle = ClassLoader::LookupTypeDefOrRefInModule(pModule,
@@ -3881,7 +3881,7 @@ PTR_CBYTE DacDbiInterfaceImpl::GetPtrToEnCField(FieldDesc * pFD, const EnCHangin
 
     PTR_EditAndContinueModule pEnCModule;
     DomainAssembly * pDomainAssembly = pEnCFieldInfo->GetObjectTypeData().vmDomainAssembly.GetDacPtr();
-    Module     * pModule     = pDomainAssembly->GetModule();
+    Module     * pModule     = pDomainAssembly->GetAssembly()->GetModule();
 
     // make sure we actually have an EditAndContinueModule
     _ASSERTE(pModule->IsEditAndContinueCapable());
@@ -4053,7 +4053,7 @@ void DacDbiInterfaceImpl::ResolveTypeReference(const TypeRefData * pTypeRefInfo,
 {
     DD_ENTER_MAY_THROW;
     DomainAssembly * pDomainAssembly        = pTypeRefInfo->vmDomainAssembly.GetDacPtr();
-    Module *     pReferencingModule = pDomainAssembly->GetModule();
+    Module *     pReferencingModule = pDomainAssembly->GetAssembly()->GetModule();
     BOOL         fSuccess = FALSE;
 
     // Resolve the type ref
@@ -4317,7 +4317,7 @@ void DacDbiInterfaceImpl::GetModuleForDomainAssembly(VMPTR_DomainAssembly vmDoma
     _ASSERTE(pModule != NULL);
 
     DomainAssembly * pDomainAssembly = vmDomainAssembly.GetDacPtr();
-    pModule->SetHostPtr(pDomainAssembly->GetModule());
+    pModule->SetHostPtr(pDomainAssembly->GetAssembly()->GetModule());
 }
 
 
@@ -4446,7 +4446,7 @@ void DacDbiInterfaceImpl::EnumerateModulesInAssembly(
     DomainAssembly * pDomainAssembly = vmAssembly.GetDacPtr();
 
     // Debugger isn't notified of Resource / Inspection-only modules.
-    if (pDomainAssembly->GetModule()->IsVisibleToDebugger())
+    if (pDomainAssembly->GetAssembly()->GetModule()->IsVisibleToDebugger())
     {
         // If domain assembly isn't yet loaded, just return
         if (!pDomainAssembly->GetAssembly()->IsLoaded())
@@ -4469,7 +4469,7 @@ VMPTR_DomainAssembly DacDbiInterfaceImpl::ResolveAssembly(
 
 
     DomainAssembly * pDomainAssembly  = vmScope.GetDacPtr();
-    Module     * pModule      = pDomainAssembly->GetModule();
+    Module     * pModule      = pDomainAssembly->GetAssembly()->GetModule();
 
     VMPTR_DomainAssembly vmDomainAssembly = VMPTR_DomainAssembly::NullPtr();
